Removing SearchContainer
