+++ src/_scss/pages/modals/extraModal/_animation.scss
@@ -13,7 +13,7 @@
 
     .eyeball-wrapper {
         position: absolute;
-        top: 15px;
+        top: 16px;
         left: 20px;
         right: 20px;
         margin-right: 20px;
